+++ packages/common/src/core/behaviors/IWithCardBehavior.ts
@@ -1,4 +1,4 @@
-import { IHubCardViewModel } from "../types/IHubCardViewModel";
+import { ICardActionLink, IHubCardViewModel } from "../types/IHubCardViewModel";
 
 export interface IWithCardBehavior {
   /**
@@ -6,10 +6,16 @@ export interface IWithCardBehavior {
    * be consumed by the suite of hub gallery components
    *
    * @param target card link contextual target
+   * @param actionLinks card action links
    * @param locale internationalization locale
    */
   convertToCardViewModel(
     target: "ago" | "view" | "workspace",
+    actionLinks: ICardActionLink[],
+    /**
+     * TODO: move transform logic to FE so we don't need to pass
+     * locale down (follow https://devtopia.esri.com/dc/hub/issues/7255)
+     */
     locale: string
   ): IHubCardViewModel;
 }