+++ packages/common/src/projects/HubProject.ts
@@ -21,7 +21,10 @@ import { ProjectEditorType } from "./_internal/ProjectSchema";
 import { IWithMetricsBehavior } from "../core/behaviors/IWithMetricsBehavior";
 import { getEntityMetrics } from "../metrics/getEntityMetrics";
 import { resolveMetric } from "../metrics/resolveMetric";
-import { IHubCardViewModel } from "../core/types/IHubCardViewModel";
+import {
+  ICardActionLink,
+  IHubCardViewModel,
+} from "../core/types/IHubCardViewModel";
 import { convertProjectEntityToCardViewModel } from "./view";
 
 /**
@@ -155,16 +158,23 @@ export class HubProject
    * be consumed by the suite of hub gallery components
    *
    * @param target card link contextual target
+   * @param actionLinks card action links
    * @param locale internationalization locale
    */
   convertToCardViewModel(
     target: "ago" | "view" | "workspace",
+    actionLinks: ICardActionLink[],
+    /**
+     * TODO: move transform logic to FE so we don't need to pass
+     * locale down (follow https://devtopia.esri.com/dc/hub/issues/7255)
+     */
     locale: string
   ): IHubCardViewModel {
     return convertProjectEntityToCardViewModel(
       this.entity,
       this.context,
       target,
+      actionLinks,
       locale
     );
   }