+++ packages/common/test/initiatives/view.test.ts
@@ -1,14 +1,18 @@
 import {
   cloneObject,
-  convertProjectEntityToCardViewModel,
-  convertProjectSearchResultToCardViewModel,
+  convertInitiativeEntityToCardViewModel,
+  convertInitiativeSearchResultToCardViewModel,
 } from "../../src";
-import { CONTEXT, PROJECT_ENTITY, PROJECT_HUB_SEARCH_RESULT } from "./fixtures";
+import {
+  CONTEXT,
+  INITIATIVE_ENTITY,
+  INITIATIVE_HUB_SEARCH_RESULT,
+} from "./fixtures";
 import * as getItemHomeUrlModule from "../../src/urls/get-item-home-url";
 import * as internalContentUtils from "../../src/content/_internal/internalContentUtils";
 import * as getRelativeWorkspaceUrlModule from "../../src/core/getRelativeWorkspaceUrl";
 
-describe("project view module:", () => {
+describe("initiative view module:", () => {
   let getShortenedCategoriesSpy: any;
 
   beforeEach(() => {
@@ -18,7 +22,7 @@ describe("project view module:", () => {
     ).and.returnValue(["category1", "category2"]);
   });
 
-  describe("convertProjectEntityToCardViewModel:", () => {
+  describe("convertInitiativeEntityToCardViewModel:", () => {
     let getItemHomeUrlSpy: any;
     let getRelativeWorkspaceUrlSpy: any;
     let getHubRelativeUrlSpy: any;
@@ -38,65 +42,65 @@ describe("project view module:", () => {
       ).and.returnValue("/mock-hub-relative-url");
     });
 
-    it("returns the card view model from the project entity", () => {
-      const result = convertProjectEntityToCardViewModel(
-        PROJECT_ENTITY,
+    it("returns the card view model from the intiative entity", () => {
+      const result = convertInitiativeEntityToCardViewModel(
+        INITIATIVE_ENTITY,
         CONTEXT
       );
 
       expect(getItemHomeUrlSpy).toHaveBeenCalledTimes(1);
       expect(getItemHomeUrlSpy).toHaveBeenCalledWith(
-        PROJECT_ENTITY.id,
+        INITIATIVE_ENTITY.id,
         CONTEXT.hubRequestOptions
       );
       expect(getHubRelativeUrlSpy).toHaveBeenCalledTimes(1);
       expect(getHubRelativeUrlSpy).toHaveBeenCalledWith(
-        PROJECT_ENTITY.type,
-        PROJECT_ENTITY.id
+        INITIATIVE_ENTITY.type,
+        INITIATIVE_ENTITY.id
       );
       expect(getRelativeWorkspaceUrlSpy).toHaveBeenCalledTimes(1);
       expect(getRelativeWorkspaceUrlSpy).toHaveBeenCalledWith(
-        PROJECT_ENTITY.type,
-        PROJECT_ENTITY.id
+        INITIATIVE_ENTITY.type,
+        INITIATIVE_ENTITY.id
       );
       expect(getShortenedCategoriesSpy).toHaveBeenCalledTimes(1);
       expect(getShortenedCategoriesSpy).toHaveBeenCalledWith(
-        PROJECT_ENTITY.categories
+        INITIATIVE_ENTITY.categories
       );
 
       expect(result).toEqual({
-        access: PROJECT_ENTITY.access,
+        access: INITIATIVE_ENTITY.access,
         actionLinks: [],
         badges: [],
-        id: PROJECT_ENTITY.id,
-        family: "project",
-        source: PROJECT_ENTITY.owner,
-        summary: PROJECT_ENTITY.summary,
-        title: PROJECT_ENTITY.name,
+        id: INITIATIVE_ENTITY.id,
+        family: "initiative",
+        source: INITIATIVE_ENTITY.owner,
+        summary: INITIATIVE_ENTITY.summary,
+        title: INITIATIVE_ENTITY.name,
         titleUrl: "https://mock-home-url.com",
-        thumbnailUrl: PROJECT_ENTITY.thumbnailUrl,
-        type: PROJECT_ENTITY.type,
+        thumbnailUrl: INITIATIVE_ENTITY.thumbnailUrl,
+        type: INITIATIVE_ENTITY.type,
         additionalInfo: [
-          { i18nKey: "type", value: PROJECT_ENTITY.type },
+          { i18nKey: "type", value: INITIATIVE_ENTITY.type },
           {
             i18nKey: "dateUpdated",
-            value: PROJECT_ENTITY.updatedDate.toLocaleDateString("en-US"),
+            value: INITIATIVE_ENTITY.updatedDate.toLocaleDateString("en-US"),
           },
-          { i18nKey: "tags", value: PROJECT_ENTITY.tags.join(", ") },
+          { i18nKey: "tags", value: INITIATIVE_ENTITY.tags.join(", ") },
           { i18nKey: "categories", value: "category1, category2" },
           {
             i18nKey: "dateCreated",
-            value: PROJECT_ENTITY.createdDate.toLocaleDateString("en-US"),
+            value: INITIATIVE_ENTITY.createdDate.toLocaleDateString("en-US"),
           },
         ],
       });
     });
     it("does not include tags/categories in the view model's additional info if none are defined", () => {
-      const modifiedEntity = cloneObject(PROJECT_ENTITY);
+      const modifiedEntity = cloneObject(INITIATIVE_ENTITY);
       modifiedEntity.tags = [];
       modifiedEntity.categories = [];
 
-      const result = convertProjectEntityToCardViewModel(
+      const result = convertInitiativeEntityToCardViewModel(
         modifiedEntity,
         CONTEXT
       );
@@ -104,16 +108,16 @@ describe("project view module:", () => {
       expect(result.additionalInfo?.length).toBe(3);
     });
     it('target = "view": returns the correct title url', () => {
-      const result = convertProjectEntityToCardViewModel(
-        PROJECT_ENTITY,
+      const result = convertInitiativeEntityToCardViewModel(
+        INITIATIVE_ENTITY,
         CONTEXT,
         "view"
       );
       expect(result.titleUrl).toBe("/mock-hub-relative-url");
     });
     it('target = "workspace": returns the correct title url', () => {
-      const result = convertProjectEntityToCardViewModel(
-        PROJECT_ENTITY,
+      const result = convertInitiativeEntityToCardViewModel(
+        INITIATIVE_ENTITY,
         CONTEXT,
         "workspace"
       );
@@ -121,59 +125,59 @@ describe("project view module:", () => {
     });
   });
 
-  describe("convertProjectSearchResultToCardViewModel", () => {
+  describe("convertInitiativeSearchResultToCardViewModel", () => {
     it("returns the card view model from the hub search result", () => {
-      const result = convertProjectSearchResultToCardViewModel(
-        PROJECT_HUB_SEARCH_RESULT
+      const result = convertInitiativeSearchResultToCardViewModel(
+        INITIATIVE_HUB_SEARCH_RESULT
       );
 
       expect(result).toEqual({
-        access: PROJECT_ENTITY.access,
+        access: INITIATIVE_ENTITY.access,
         actionLinks: [],
         badges: [],
-        id: PROJECT_ENTITY.id,
-        family: "project",
-        source: PROJECT_ENTITY.owner,
-        summary: PROJECT_ENTITY.summary,
-        title: PROJECT_ENTITY.name,
+        id: INITIATIVE_ENTITY.id,
+        family: "initiative",
+        source: INITIATIVE_ENTITY.owner,
+        summary: INITIATIVE_ENTITY.summary,
+        title: INITIATIVE_ENTITY.name,
         titleUrl: "https://mock-home-url.com",
-        thumbnailUrl: PROJECT_ENTITY.thumbnailUrl,
-        type: PROJECT_ENTITY.type,
+        thumbnailUrl: INITIATIVE_ENTITY.thumbnailUrl,
+        type: INITIATIVE_ENTITY.type,
         additionalInfo: [
-          { i18nKey: "type", value: PROJECT_ENTITY.type },
+          { i18nKey: "type", value: INITIATIVE_ENTITY.type },
           {
             i18nKey: "dateUpdated",
-            value: PROJECT_ENTITY.updatedDate.toLocaleDateString("en-US"),
+            value: INITIATIVE_ENTITY.updatedDate.toLocaleDateString("en-US"),
           },
-          { i18nKey: "tags", value: PROJECT_ENTITY.tags.join(", ") },
+          { i18nKey: "tags", value: INITIATIVE_ENTITY.tags.join(", ") },
           { i18nKey: "categories", value: "category1, category2" },
           {
             i18nKey: "dateCreated",
-            value: PROJECT_ENTITY.createdDate.toLocaleDateString("en-US"),
+            value: INITIATIVE_ENTITY.createdDate.toLocaleDateString("en-US"),
           },
         ],
       });
     });
     it("does not include tags/categories in the view model's additional info if none are defined", () => {
-      const modifiedSearchResult = cloneObject(PROJECT_HUB_SEARCH_RESULT);
+      const modifiedSearchResult = cloneObject(INITIATIVE_HUB_SEARCH_RESULT);
       modifiedSearchResult.tags = undefined;
       modifiedSearchResult.categories = undefined;
 
       const result =
-        convertProjectSearchResultToCardViewModel(modifiedSearchResult);
+        convertInitiativeSearchResultToCardViewModel(modifiedSearchResult);
 
       expect(result.additionalInfo?.length).toBe(3);
     });
     it('target = "view": returns the correct title url', () => {
-      const result = convertProjectSearchResultToCardViewModel(
-        PROJECT_HUB_SEARCH_RESULT,
+      const result = convertInitiativeSearchResultToCardViewModel(
+        INITIATIVE_HUB_SEARCH_RESULT,
         "view"
       );
       expect(result.titleUrl).toBe("/mock-hub-relative-url");
     });
     it('target = "workspace": returns the correct title url', () => {
-      const result = convertProjectSearchResultToCardViewModel(
-        PROJECT_HUB_SEARCH_RESULT,
+      const result = convertInitiativeSearchResultToCardViewModel(
+        INITIATIVE_HUB_SEARCH_RESULT,
         "workspace"
       );
       expect(result.titleUrl).toBe("/mock-relative-workspace-url");